backend/requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
Flask-Swagger-UI==4.11.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.27

# Redis
redis==5.0.1

# ATT&CK Integration
mitreattack-python==2.0.10
stix2==3.0.1
taxii2-client==2.3.0

# Document Generation
reportlab==4.0.8
python-docx==0.8.11
openpyxl==3.1.2
Pillow==10.2.0

# Task Queue
celery==5.3.6
flower==2.0.1

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
python-multipart==0.0.6
requests==2.31.0
urllib3==2.2.0
certifi==2024.2.2
chardet==5.2.0
idna==3.6
jsonschema==4.21.1

# Logging
python-json-logger==2.0.7
structlog==24.1.0

# Dev Tools
pytest==7.4.4
pytest-flask==1.2.0
pytest-cov==4.1.0
flake8==7.0.0
black==24.1.0
isort==5.13.2

# Production
gunicorn==21.2.0
eventlet==0.33.3
gevent==23.7.0
